pyproject.toml
[tool.black]
line-length = 88
py36 = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.tox
  | venv
  | _build
  | buck-out
  | build
  | dist
  | docs
)/
'''

[tool.poetry]
name = "graphene-django-flufy"
version = "0.1.0"
description = "This library add some extra funcionalities to graphene-django to facilitate the graphql use without Relay, allow paginations and filtering integration and add some extra directives"
authors = ["Daniel Ale <danielale9291@gmail.com>"]
readme = "README.md"
packages = [{include = "graphene_django_flufy"}]
license = "MIT"
homepage="https://github.com/profmcdan/graphene-django-flufy"
repository="https://github.com/profmcdan/graphene-django-flufy"
documentation="https://github.com/profmcdan/graphene-django-flufy"
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: PyPy",
]
keywords=["api", "graphql", "protocol", "graphene", "django"]

[tool.poetry.dependencies]
python = "^3.7 || ^3.8 || ^3.9 || ^3.10"
graphene-django = "^3.0.0"
python-dateutil = "^2.8.2"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
rx = "^3.2.0"
graphql-core = "^3.2.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = "^22.10.0"
tox = "^3.27.0"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
ipdb = "^0.13.9"
flake8 = "^5.0.4"
bandit = "^1.7.4"
pytest-cov = "^4.0.0"
pyflakes = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
